Simplify grenache network launching
