+++ src/helpers/index.js
@@ -1,10 +1,6 @@
 'use strict'
 
 const {
-  bootTwoGrapes,
-  killGrapes
-} = require('./grapes')
-const {
   getFreePort,
   getDefaultPorts
 } = require('./ports')
@@ -13,8 +9,6 @@ const {
 } = require('./utils')
 
 module.exports = {
-  bootTwoGrapes,
-  killGrapes,
   getFreePort,
   getDefaultPorts,
   checkAndChangeAccess
